Dirty: Emit version rather than publish, move logout on version change logic to socketWithPresence decorator
